# be/openapi.yaml
openapi: 3.0.3
info:
  title: Dietonez
  version: 1.0.0
  description: My vibe-coded tool for diet planning.
servers:
  - url: http://:192.46.236.119/api/v1
    description: Main API server
paths:
...
  /dishes:
    get:
      summary: Get all dishes of a given meal category
      description: Returns a list of dishes for the specified meal type, sorted alphabetically by name.
      operationId: getDishes
      tags:
        - Dishes
      parameters:
        - name: meal
          in: query
          required: true
          description: Meal type to filter dishes (e.g., Breakfast, MainMeal, PreWorkout, Supper)
          schema:
            type: string
            enum: [Breakfast, MainMeal, PreWorkout, Supper]
        - name: min
          in: query
          required: false
          description: Return minimal dish information (ID and name only)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishGetShort'
    post:
      summary: Create a new dish
      description: Creates a new dish based on input and returns the full object with computed macros and assigned ID.
      operationId: createDish
      tags:
        - Dishes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/DishPost'
              required:
                - name
                - meal
                - ingredients
                - recipe
              properties:
                id:
                  readOnly: true
      responses:
        '200':
          description: Created dish with computed macros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishGet'
        '400':
          description: Invalid input or missing fields
  /dishes/{id}:
    get:
      summary: Get dish by ID
      description: Returns full details of a single dish, including ingredients and computed macros.
      operationId: getDishById
      tags:
        - Dishes
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the dish
          schema:
            type: integer
      responses:
        '200':
          description: Dish object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishGet'
        '404':
          description: Dish not found
    put:
      summary: Update an existing dish
      description: Updates an existing dish by ID and returns the updated object with recomputed macros.
      operationId: updateDish
      tags:
        - Dishes
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the dish to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishPut'
      responses:
        '200':
          description: Updated dish with recomputed macros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishGet'
        '400':
          description: Invalid input or bad format
        '404':
          description: Dish not found
    delete:
      summary: Delete dish by ID
      description: Deletes a dish. Returns 204 if successful, or 400 if the dish is used in a diet.
      operationId: deleteDish
      tags:
        - Dishes
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the dish to delete
          schema:
            type: integer
      responses:
        '204':
          description: Dish deleted successfully
        '400':
          description: Dish is used in a diet and cannot be deleted
...
components:
  schemas:
    IngredientGetPut:
      type: object
      properties:
        id:
          type: integer
          example: 0
        name:
          type: string
          example: "Rolada ustrzycka (Regionalne szlaki)"           
        kcal:
          type: number
          example: 543
        protein:
          type: number
          example: 28
        fat:
          type: number
          example: 28
        carbs:
          type: number
          example: 0
        unit:
          type: string
          example: "g"
        shopStyle:
          type: string
          example: "Lidl"
        default_amount:
          type: number
          example: 100
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    IngredientMin:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Rolada Ustrzycka (Regionalne Szlaki)"
...
    DishGetShort:
      type: object
      required: [id, name, kcal, protein, fat, carbs]
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: "Tofu Curry"
        kcal:
          type: number
          format: float
          example: 620.5
        protein:
          type: number
          format: float
          example: 40.1
        fat:
          type: number
          format: float
          example: 25.3
        carbs:
          type: number
          format: float
          example: 50.0
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    Label:
      type: object
      required: [text, color]
      properties:
        text:
          type: string
          example: "Wege"
        color:
          type: string
          format: hex
          example: "#4caf50"
    DishGet:
      type: object
      required:
        - id
        - name
        - descr
        - meal
        - kcal
        - protein
        - fat
        - carbs
        - ingredients
        - recipe
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "M1 Kurczak Słodko Kwaśny"
        meal:
          type: string
          enum: [Breakfast, MainMeal, PreWorkout, Supper]
          example: "Supper"
        kcal:
          type: number
          example: 573
        protein:
          type: number
          example: 34
        fat:
          type: number
          example: 24
        carbs:
          type: number
          example: 78
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientInDishGet'
        recipe:
          $ref: '#/components/schemas/Recipe'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    DishPost:
      type: object
      required: [name, meal, ingredients, recipe, labels]
      properties:
        name:
          type: string
          example: "M2 Wołowina - burger"
        meal:
          type: string
          enum: [Breakfast, MainMeal, PreWorkout, Supper]
          example: "MainMeal"
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientInDishPut'
        recipe:
          $ref: '#/components/schemas/Recipe'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    DishPut:
      type: object
      required: [id, name, meal, ingredients, recipe, labels]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "M3 Dorsz - filet z frytkami"
        meal:
          type: string
          enum: [Breakfast, MainMeal, PreWorkout, Supper]
          example: "PreWorkout"
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientInDishPut'
        recipe:
          $ref: '#/components/schemas/Recipe'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
...
    IngredientInDishGet:
      type: object
      required: [ingredient, amount]
      properties:
        ingredient:
          $ref: '#/components/schemas/IngredientGetPut'
        amount:
          type: number
          format: float
          example: 12
    IngredientInDishPut:
      type: object
      required: [ingredient, amount]
      properties:
        ingredient:
          $ref: '#/components/schemas/IngredientMin'
        amount:
          type: number
          format: float
          example: 1.5
    Recipe:
      type: object
      required: [total_time, before, when_to_start, preparation]
      properties:
        total_time:
          type: string
          example: "ok. 20 minut"
        before:
          type: string
          example: "nie wymaga wcześniejszego przygotowania."
        when_to_start:
          type: string
          example: "ok. 20 minut przed posiłkiem"
        preparation:
          type: string
          example: "Kurczaka pokrój w paski i zamarynuj w oliwie z przyprawami. Smaż 8–10 minut"
...